+++ pytsite/package_info/__init__.py
@@ -1,7 +1,7 @@
 """PytSite Package Utilities
 """
 from . import _error as error
-from ._api import resolve_path, parse_json, data, requires, requires_packages, requires_plugins, name, description, \
+from ._api import resolve_package_path, parse_json, data, requires, requires_packages, requires_plugins, name, description, \
     version, url
 
 __author__ = 'Alexander Shepetko'