+++ pytsite/plugman/__init__.py
@@ -5,14 +5,31 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 # Public API
+from sys import meta_path as _meta_path
+from os import path as _path
 from . import _error as error
 from ._api import plugins_path, plugin_package_info, install, uninstall, is_installed, load, is_loaded, \
     plugins_info, remote_plugin_info, remote_plugins_info, is_dev_mode, get_dependant_plugins, \
-    get_allowed_version_range, on_install, on_uninstall
+    get_allowed_version_range, on_install, on_uninstall, plugin_path, is_loading
+
+
+class _MetaPathHook:
+    def find_spec(self, name: str, import_path: list, module=None):
+        name_s = name.split('.')
+        if name_s[0] == 'plugins' and len(name_s) == 2:
+            # Check if the plugin installed
+            p_path = plugin_path(name_s[1])
+            search_for = p_path if is_dev_mode() else _path.join(p_path, 'installed')
+            if not _path.exists(search_for):
+                raise error.PluginNotInstalled(name_s[1])
+
+            # Check if the plugin loaded
+            if not is_loaded(name_s[1]) and not is_loading(name_s[1]):
+                raise error.PluginNotLoaded(name_s[1])
 
 
 def _init():
-    from os import mkdir, path
+    from os import mkdir
     from pytsite import lang, console, update
     from . import _console_command
 
@@ -21,9 +38,9 @@ def _init():
 
     # Create 'plugins' package if it doesn't exist
     plugins_dir_path = plugins_path()
-    if not path.exists(plugins_dir_path):
+    if not _path.exists(plugins_dir_path):
         mkdir(plugins_dir_path, 0o755)
-        with open(path.join(plugins_dir_path, '__init__.py'), 'wt') as f:
+        with open(_path.join(plugins_dir_path, '__init__.py'), 'wt') as f:
             f.write('"""Pytsite Plugins\n"""\n')
 
     # Register console commands
@@ -31,8 +48,14 @@ def _init():
     console.register_command(_console_command.Update())
     console.register_command(_console_command.Uninstall())
 
+    # Get information about installed plugins
+    plugins_info_seq = plugins_info()
+
+    # Enable imports checking
+    _meta_path.insert(0, _MetaPathHook())
+
     # Load installed plugins
-    for p_name in plugins_info():
+    for p_name in plugins_info_seq:
         if p_name.startswith('_'):
             continue