+++ pytsite/router/__init__.py
@@ -1,18 +1,19 @@
 """PytSite Router
 """
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
 from ._api import handle, add_path_alias, base_path, base_url, call, current_path, current_url, dispatch, rule_path, \
     rule_url, is_base_url, has_rule, get_no_cache, set_no_cache, remove_path_alias, scheme, server_name, url, \
     session, request, set_request, get_session_store, on_pre_dispatch, on_dispatch, on_response, on_exception, \
     on_xhr_dispatch, on_xhr_pre_dispatch, on_xhr_response, delete_session
 
-__author__ = 'Alexander Shepetko'
-__email__ = 'a@shepetko.com'
-__license__ = 'MIT'
-
 
 def _init():
     from os import path, makedirs
-    from pytsite import tpl, lang, reg
+    from pytsite import tpl, lang, reg, cleanup
+    from . import _eh
 
     # Resources
     tpl.register_package(__name__)
@@ -35,10 +36,13 @@ def _init():
     tpl.register_global('is_base_url', is_base_url)
     tpl.register_global('session_messages', lambda x: session().get_messages(x) if session() else ())
 
-    # Clear flash messages from all session
+    # Clear flash messages from all sessions
     s_store = get_session_store()
     for sid in s_store.list():
         s_store.save_if_modified(s_store.get(sid).flash_clear())
+
+    # Events handlers
+    cleanup.on_cleanup(_eh.cleanup)
 
 
 _init()