+++ pytsite/util/__init__.py
@@ -1,15 +1,15 @@
-"""PytSite Helpers.
+"""PytSite Helpers
 """
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
 # Public API
 from werkzeug.urls import url_quote, url_unquote
 from . import _error as error
 from ._api import strip_html_tags, cleanup_dict, cleanup_list, dict_merge, escape_html, get_module_attr, html_attrs_str, \
     md5_hex_digest, mk_tmp_file, nav_link, random_password, random_str, rfc822_datetime_str, transform_str_1, \
     transform_str_2, transliterate, trim_str, w3c_datetime_str, weight_sort, parse_rfc822_datetime_str, \
-    parse_w3c_datetime_str, minify_html, to_snake_case, tidyfy_html, format_call_stack_str, mk_tmp_dir, is_url, \
+    parse_w3c_datetime_str, minify_html, to_snake_case, tidyfy_html, get_call_stack, mk_tmp_dir, is_url, \
     load_json, install_pip_package, get_installed_pip_package_info, get_installed_pip_package_version, \
-    is_pip_package_installed
-
-__author__ = 'Alexander Shepetko'
-__email__ = 'a@shepetko.com'
-__license__ = 'MIT'
+    is_pip_package_installed, remove_obsolete_files